+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheetFormFactory/PaymentSheetFormFactory.swift
@@ -153,12 +153,21 @@ class PaymentSheetFormFactory {
         } else if paymentMethod.stpPaymentMethodType == .payPal && saveMode == .merchantRequired {
             // Paypal requires mandate when setting up
             additionalElements = [makePaypalMandate()]
+        } else if paymentMethod.stpPaymentMethodType == .bancontact {
+            return makeBancontact()
+        }
+
+        guard let spec = specFromJSONProvider() else {
+            assertionFailure("Failed to get form spec!")
+            return FormElement(elements: [], theme: theme)
+        }
+        if paymentMethod.stpPaymentMethodType == .iDEAL {
+            return makeiDEAL(spec: spec)
+        } else if paymentMethod.stpPaymentMethodType == .sofort {
+            return makeSofort(spec: spec)
         }
 
         // 2. Element-based forms defined in JSON
-        guard let spec = specFromJSONProvider() else {
-            fatalError()
-        }
         return makeFormElementFromSpec(spec: spec, additionalElements: additionalElements)
     }
 }
@@ -325,7 +334,8 @@ extension PaymentSheetFormFactory {
 
     func makeBillingAddressSection(
         collectionMode: AddressSectionElement.CollectionMode = .all(),
-        countries: [String]?
+        countries: [String]? = nil,
+        countryAPIPath: String? = nil
     ) -> PaymentMethodElementWrapper<AddressSectionElement> {
         let displayBillingSameAsShippingCheckbox: Bool
         let defaultAddress: AddressSectionElement.AddressDetails
@@ -371,12 +381,83 @@ extension PaymentSheetFormFactory {
                 params.paymentMethodParams.nonnil_billingDetails.nonnil_address.postalCode = postalCode.text
             }
             params.paymentMethodParams.nonnil_billingDetails.nonnil_address.country = section.selectedCountryCode
+            if let countryAPIPath {
+                params.paymentMethodParams.additionalAPIParameters[countryAPIPath] = section.selectedCountryCode
+            }
 
             return params
         }
     }
 
     // MARK: - PaymentMethod form definitions
+
+    func makeSofort(spec: FormSpec) -> PaymentMethodElement {
+        let contactSection: Element? = makeContactInformationSection(
+            nameRequiredByPaymentMethod: saveMode == .merchantRequired,
+            emailRequiredByPaymentMethod: saveMode == .merchantRequired,
+            phoneRequiredByPaymentMethod: false
+        )
+        // Hack: Use the luxe spec to get the latest list of accepted countries rather than hardcoding it here
+        let countries: [String]? = spec.fields.reduce(nil) { countries, fieldSpec in
+            if case let .country(countrySpec) = fieldSpec {
+                return countrySpec.allowedCountryCodes
+            }
+            return countries
+        }
+
+        let addressSection: Element? = {
+            if configuration.billingDetailsCollectionConfiguration.address == .full {
+                return makeBillingAddressSection(countries: countries, countryAPIPath: "sofort[country]")
+            } else {
+                return makeCountry(countryCodes: countries, apiPath: "sofort[country]")
+            }
+        }()
+        let mandate: Element? = saveMode == .merchantRequired ? makeSepaMandate() : nil // Note: We show a SEPA mandate b/c sofort saves bank details as a SEPA Direct Debit Payment Method
+        let elements: [Element?] = [contactSection, addressSection, mandate]
+        return FormElement(
+            autoSectioningElements: elements.compactMap { $0 },
+            theme: theme
+        )
+    }
+
+    func makeBancontact() -> PaymentMethodElement {
+        let contactSection: Element? = makeContactInformationSection(
+            nameRequiredByPaymentMethod: true,
+            emailRequiredByPaymentMethod: saveMode == .merchantRequired,
+            phoneRequiredByPaymentMethod: false
+        )
+        let addressSection: Element? = makeBillingAddressSectionIfNecessary(requiredByPaymentMethod: false)
+        let mandate: Element? = saveMode == .merchantRequired ? makeSepaMandate() : nil // Note: We show a SEPA mandate b/c iDEAL saves bank details as a SEPA Direct Debit Payment Method
+        let elements: [Element?] = [contactSection, addressSection, mandate]
+        return FormElement(
+            autoSectioningElements: elements.compactMap { $0 },
+            theme: theme
+        )
+    }
+
+    func makeiDEAL(spec: FormSpec) -> PaymentMethodElement {
+        let contactSection: Element? = makeContactInformationSection(
+            nameRequiredByPaymentMethod: true,
+            emailRequiredByPaymentMethod: saveMode == .merchantRequired,
+            phoneRequiredByPaymentMethod: false
+        )
+        // Hack: Use the luxe spec to make the dropdown for convenience; it has the latest list of banks
+        let bankDropdown: Element? = spec.fields.reduce(nil) { dropdown, spec in
+            // Find the dropdown spec
+            if case .selector(let spec) = spec {
+                return makeDropdown(for: spec)
+            }
+            return dropdown
+        }
+
+        let addressSection: Element? = makeBillingAddressSectionIfNecessary(requiredByPaymentMethod: false)
+        let mandate: Element? = saveMode == .merchantRequired ? makeSepaMandate() : nil // Note: We show a SEPA mandate b/c iDEAL saves bank details as a SEPA Direct Debit Payment Method
+        let elements: [Element?] = [contactSection, bankDropdown, addressSection, mandate]
+        return FormElement(
+            autoSectioningElements: elements.compactMap { $0 },
+            theme: theme
+        )
+    }
 
     func makeUSBankAccount(merchantName: String) -> PaymentMethodElement {
         let isSaving = BoolReference()
@@ -556,9 +637,18 @@ extension PaymentSheetFormFactory {
         guard !elements.isEmpty else { return nil }
 
         return SectionElement(
-            title: .Localized.contact_information,
+            title: elements.count > 1 ? .Localized.contact_information : nil,
             elements: elements,
             theme: theme)
+    }
+
+    func makeBillingAddressSectionIfNecessary(requiredByPaymentMethod: Bool) -> Element? {
+        if configuration.billingDetailsCollectionConfiguration.address == .full
+            || (configuration.billingDetailsCollectionConfiguration.address == .automatic && requiredByPaymentMethod) {
+           return makeBillingAddressSection()
+        } else {
+            return nil
+        }
     }
 
     func makeDefaultsApplierWrapper<T: PaymentMethodElement>(for element: T) -> PaymentMethodElementWrapper<T> {